move class - Role
