+++ src/main/java/co/fineants/api/domain/member/domain/factory/MemberRoleFactory.java
@@ -2,8 +2,8 @@ package co.fineants.api.domain.member.domain.factory;
 
 import co.fineants.api.domain.member.domain.entity.Member;
 import co.fineants.api.domain.member.domain.entity.MemberRole;
-import co.fineants.api.domain.member.domain.entity.Role;
 import co.fineants.api.domain.member.service.RoleService;
+import co.fineants.api.domain.role.domain.Role;
 
 public class MemberRoleFactory {
