+++ src/main/java/co/fineants/api/domain/member/service/RoleService.java
@@ -3,8 +3,8 @@ package co.fineants.api.domain.member.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import co.fineants.api.domain.member.domain.entity.Role;
 import co.fineants.api.domain.member.repository.RoleRepository;
+import co.fineants.api.domain.role.domain.Role;
 import co.fineants.api.global.errors.exception.business.RoleNotFoundException;
 
 @Service
